service/requirements.txt
fastapi
dill
numpy
uvicorn
pydantic
pandas
gensim
ipython
Jinja2
kolibri-ml
seaborn
dtreeviz
simplejson